fix(table): move common test util out of table
